feat: refactor device communication by introducing DeviceCommunicator class
