+++ GreeMQTT/device_communication.py
@@ -1,15 +1,15 @@
 import asyncio
+from typing import Optional
 
-SOCKET_TIMEOUT = 5
-BUFFER_SIZE = 1024
 UDP_PORT = 7000
+SOCKET_TIMEOUT = 5
 
 
-class DeviceCommunication:
+class DeviceCommunicator:
     def __init__(self, device_ip: str):
         self.device_ip = device_ip
 
-    async def send_data(self, request: bytes) -> bytes | None:
+    async def send_data(self, request: bytes) -> Optional[bytes]:
         loop = asyncio.get_running_loop()
         on_con_lost = loop.create_future()
         response_data = bytearray()
@@ -35,6 +35,40 @@ class DeviceCommunication:
         transport, protocol = await loop.create_datagram_endpoint(
             lambda: UDPClientProtocol(self.device_ip),
             remote_addr=(self.device_ip, UDP_PORT),
+        )
+        try:
+            await asyncio.wait_for(on_con_lost, timeout=SOCKET_TIMEOUT)
+        except asyncio.TimeoutError:
+            transport.close()
+            return None
+        return bytes(response_data) if response_data else None
+
+    @staticmethod
+    async def broadcast_scan(ip_address: str) -> Optional[bytes]:
+        loop = asyncio.get_running_loop()
+        on_con_lost = loop.create_future()
+        response_data = bytearray()
+
+        class UDPBroadcastProtocol(asyncio.DatagramProtocol):
+            def __init__(self):
+                self.transport = None
+
+            def connection_made(self, transport):
+                self.transport = transport
+                self.transport.sendto(b'{"t":"scan"}', (ip_address, UDP_PORT))
+
+            def datagram_received(self, data, addr):
+                response_data.extend(data)
+                on_con_lost.set_result(True)
+                self.transport.close()
+
+            def error_received(self, exc):
+                on_con_lost.set_result(False)
+                self.transport.close()
+
+        transport, protocol = await loop.create_datagram_endpoint(
+            lambda: UDPBroadcastProtocol(),
+            remote_addr=(ip_address, UDP_PORT),
         )
         try:
             await asyncio.wait_for(on_con_lost, timeout=SOCKET_TIMEOUT)
